refactor: playground 패키지 정리
