+++ playground/src/main/kotlin/team/duckie/quackquack/playground/base/BaseActivity.kt
@@ -7,7 +7,7 @@
  * Please see full license: https://github.com/sungbinland/quack-quack/blob/main/LICENSE
  */
 
-package team.duckie.quackquack.playground
+package team.duckie.quackquack.playground.base
 
 import android.content.res.Configuration
 import android.graphics.Color
